// src/app/models/types/Proposal/ProposalModel.ts
export type ProposalModel = {
    SNO?: number;
    UserId?: number;
    ProposalId?: number;
    ProposalNo?: string;
    DealerId?: number;
    CreatedDate?: string;
    ChassisNo?: string;
    VehRegistionNo?: string;
    ModelVariantName?: string;
    PolicyType?: string;
    ICName?: string;
    RenewalId?: number;
    CustomerName?: string;
}